% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{comparedObj-class}
\alias{comparedObj-class}
\alias{comparedObj}
\title{S4 class defining comparedObj class}
\value{
comparedObj
}
\description{
\code{comparedObj} defines compareNets outputs.
}
\section{Slots}{

\describe{
\item{\code{ji}}{The Jaccard index between grph.1 and grph.2,
considering both shared genes and edges.}

\item{\code{permutations}}{A numeric vector of calculated Jaccard indices,
comparing pairs of randomly reconstructed networks that 
are sampled from a union of observed grph.1 and grph.2.}

\item{\code{pvalue}}{The \code{permutations} distribution is used to calculate a 
two-sided t-test p-value, which reflects the proportion 
of permuted Jaccard indices that differ from the observed \code{ji}.}

\item{\code{deltas_df}}{A data frame containing the features calculated per removal 
                     of a given gene by comparing two networks grph.1 and grph.2.
\itemize{ 
          \item    \code{gene}: gene name. 
          \item    \code{jaccard}: The Jaccard similarity index between grph.1 and grph.2, 
          after gene removal.
          \item    \code{degree.1} and \code{degree.2}: the normalized degree of a
          given gene in grph.1 and grph.2. Zero, if the gene does not exist.
          \item    \code{closeness.1} and \code{closeness.2}: the normalized closeness
          of a given gene in grph.1 and grph.2. Zero, if the gene does not
          exist.
          \item    \code{betweenness.1} and \code{betweenness.2}: the normalized betweenness
          of a given gene in grph.1 and grph.2. Zero, if the gene does not exist.
          \item    \code{delta_jaccard}: the relative difference in similarity of 
          grph.1 and grph.2 after removal of given gene (\code{jaccard}) compared 
          to before removal (\code{ji}).
          \item    \code{delta_degree}: the difference between degree.1 and degree.2.
          \item    \code{delta_closeness}: the difference between closeness.1 and closeness.2
          \item    \code{delta_betweenness}: the difference between betweenness.1 and betweenness.2.
          \item    \code{type}: wether a gene is hub in grph.1, grph.2, both, or non.
          }}

\item{\code{inA}}{genes in grph.1 and not in grph.2}

\item{\code{inB}}{genes in grph.2 and not in grph.1}

\item{\code{inAB}}{genes in both grph.1 and grph.2}

\item{\code{inAB_mtx}}{table of gene counts in grph.1 and grph.2}
}}

