% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{compiledObj-class}
\alias{compiledObj-class}
\alias{compiledObj}
\title{S4 class defining compiledObj class}
\value{
compiledObj
}
\description{
\code{compiledObj} defines compileNets outputs.
}
\section{Slots}{

\describe{
\item{\code{graph}}{a compiled estimated inverse covariance graph. The edges indicate
the partial correlation values. The nodes represent the genes.}

\item{\code{graph_df}}{converted graph to a long data frame:
\itemize{ 
          \item    \code{from_gene}: vertex A. 
          \item    \code{from_connectivity}: vertex A degree. 
          \item    \code{from_module}: the module vertex A belongs to. 
          \item    \code{from_hub}: is the vertex A, a hub gene. 
          \item    \code{par_corr_coeff}: The weight of given edge equal to 
                                          the partial correlation coefficient. 
          \item    \code{to_gene}: vertex B. 
          \item    \code{to_connectivity}: vertex B degree. 
          \item    \code{to_module}: the module vertex B belongs to. 
          \item    \code{to_hub}: is the vertex B, a hub gene. 
          }}

\item{\code{hubs}}{A named vector containing the names of hub genes
and their connectivity, sorted in ascending order.}

\item{\code{modules}}{A named list containing the module gene names and
their connectivity, sorted from lowest degree 
ascending. The modules themselves are sorted in descending 
order by size, starting from the largest module (M0).}
}}

