% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{constructNets}
\alias{constructNets}
\title{Construct gene association network}
\usage{
constructNets(
  gex,
  glasso.rho = 1,
  metadata = NULL,
  field = NULL,
  scale.gex = FALSE,
  qc.gex = c(0, 0),
  ...
)
}
\arguments{
\item{gex}{A numeric matrix of single-cell expression values
where rows are genes and columns are cells.}

\item{glasso.rho}{Regularization parameter for lasso. This scalar value 
controls the degree of sparsity in the resulting
graph. Higher values result in greater sparsity. 
See \link[glasso]{glasso}. Default is 1.}

\item{metadata}{Provide to split the input \code{gex} matrix by 
\code{field}. Row names should represent cell barcodes 
similar to the column names of the \code{gex} matrix.}

\item{field}{The column name from \code{metadata} by which to
split the input \code{gex} matrix. This parameter 
results in one graph per group. Default is no spliting.}

\item{scale.gex}{if TRUE, the count expression matrix is normalized 
to the mean library size, where each cell is scaled 
to sum up to the mean total counts. Default is false.}

\item{qc.gex}{A numeric vector of two fractions. The first fraction 
is used as a threshold to filter out cells that express 
fewer number of genes than this fraction. The second fraction is 
employed to exclude genes expressed in fewer number of cells than 
this specified fraction. Default is a minimum qc to remove 
zero rows and columns.}

\item{...}{Additional arguments to be passed to \link[glasso]{glasso}. 
Otherwise, default values from the \code{glasso} 
function are used.}
}
\value{
a \link{genixObj} object. It can be a list of \code{genixObj}s if 
          \code{metadata} and \code{field} are provided.
}
\description{
A glasso-based algorithm to infer undirected dependency graphs 
               from scRNA-seq data
}
\details{
\code{constructNets} uses a glasso-based algorith to infer gene association 
         networks from a scRNA-seq data.
}
\examples{
\dontrun{
# getting the data:
data("genixDemoData")

# construct a graph per group:
rslts <- constructNets(gex=genixDemoData$gex, glasso.rho=1.0, 
                       metadata=genixDemoData$metadata, 
                       field="group", scale.gex=TRUE, 
                       qc.gex=c(0.01,0.5), maxit=1e3)
}

}
\seealso{
\link{genixObj} for a genix object. \link[glasso]{glasso} for 
          details about glasso algorith.
}
