% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{fetchTVGs}
\alias{fetchTVGs}
\title{Topological variant gene identification}
\usage{
fetchTVGs(obj, feature, sigma = 1)
}
\arguments{
\item{obj}{a \link{comparedObj} object.}

\item{feature}{a character. The feature by which to filter the data.}

\item{sigma}{a numeric. The standard deviation coefficient by which to 
filter the data.}
}
\value{
A data.frame containing gene names that have passed the filter of
  \code{sigma}*\code{sd}(\code{feature}). The output data.frame is sorted by
  the values of \code{feature} in descending order.
}
\description{
filter genes with topological feastures above a threshold.
}
\examples{
\dontrun{
# first run \link{compareNets} function example and pass the \code{cmpr_rslts} 

x <- fetchTVGs(cmpr_rslts, feature="delta_degree", sigma=1)
}

}
