% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genix.R
\docType{package}
\name{genix}
\alias{genix}
\title{The genix package}
\description{
\code{genix} provides a foundation for sinature discovery through
comparative analysis of gene association networks using single cell RNA
sequencing data. For additional details regarding the use of the \code{genix}
package see the vignettes:\cr
\code{browseVignettes("genix")}
}
\section{Construction, Process, and Comparison of Gene Association Networks}{

\itemize{
  \item  \link{constructNets}:     Construct gene association network from
                                  scRNA-seq data.
  \item  \link{compileNets}:       Compile the constructed network and
                                  identify hub gene and gene modules.
  \item  \link{compareNets}:       Compare two ntworks and report topological 
                                  variant genes.
}
}

\references{
\enumerate{
  \item  Nima Nouri, et al.: Signature Genes Identification through
                             Comparative Analysis of Gene Association
                             Networks Using Single-Cell RNA Sequencing Data
}
}
