% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{genixDemoData}
\alias{genixDemoData}
\title{Example for demo}
\format{
An object with the following items:
  \itemize{
    \item  \code{gex}:               numeric matrix of single-cell 
                                     expression values where rows are 
                                     genes and columns are cells.
    \item  \code{metadata}:          data.frame of meta data including 
                                     typical \code{Seurat} metadata
                                     columns such as \code{orig.ident},
                                     \code{nCount_RNA}, and \code{nFeature_RNA}.
                                     Additionally, it includes a \code{group}
                                     column that identifies the group to
                                     which each cell belongs.
}
}
\usage{
data("genixDemoData")
}
\value{
An R object.
}
\description{
A small example database containing the gene expression matrix of 200 cells.
}
\keyword{datasets}
