% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{genixObj-class}
\alias{genixObj-class}
\alias{genixObj}
\title{S4 class defining genixObj class}
\value{
genixObj
}
\description{
\code{genixObj} defines constructNets outputs.
}
\section{Slots}{

\describe{
\item{\code{icm_graph}}{an estimated inverse covariance graph. The edges indicate
the partial correlation values. The nodes represent the genes.}
}}

