% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{plotNetDegree}
\alias{plotNetDegree}
\title{plots inferred network characteristics}
\usage{
plotNetDegree(obj)
}
\arguments{
\item{obj}{a \link{compiledObj} object.}
}
\value{
A ggplot object.
}
\description{
Displays the distribution of degrees in a graph.
}
\examples{
\dontrun{
# first run  \link{compileNets} function example and pass the \link{cmpl_rslts} 

# plot degree distribution:
plotNetDegree(cmpl_rslts$grp2)
}

}
