% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{plotPermutes}
\alias{plotPermutes}
\title{graphs distinctiveness}
\usage{
plotPermutes(obj)
}
\arguments{
\item{obj}{an \link{comparedObj} object.}
}
\value{
A ggplot object.
}
\description{
Displays the distribution of Jaccard index similarities
}
\examples{
\dontrun{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)

fetchDotPlot(srgnt, 
             genes = c("CD3D","CD3E","IL7R","ANPEP","SLC13A3","SLC17A3",
                       "EMCN","NOTCH4","ADGRL4","CD79A","MS4A1","CD19",
                       "CD14","MS4A7","FCGR3A"), 
             min.pct=0.1)
}

}
\seealso{
\code{compareNets} to generate \link{comparedObj}.
}
