% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{plotNetDegree}
\alias{plotNetDegree}
\title{plots inferred network characteristics}
\usage{
plotNetDegree(grph)
}
\arguments{
\item{grph}{an igraph object.}
}
\value{
A ggplot object.
}
\description{
Displays the distribution of degrees in a graph.
}
\examples{
\dontrun{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)

fetchDotPlot(srgnt, 
             genes = c("CD3D","CD3E","IL7R","ANPEP","SLC13A3","SLC17A3",
                       "EMCN","NOTCH4","ADGRL4","CD79A","MS4A1","CD19",
                       "CD14","MS4A7","FCGR3A"), 
             min.pct=0.1)
}

}
