% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{plotPermutes}
\alias{plotPermutes}
\title{graphs distinctiveness}
\usage{
plotPermutes(obj)
}
\arguments{
\item{obj}{a \link{comparedObj} object.}
}
\value{
A ggplot object.
}
\description{
Displays the distribution of Jaccard index similarities
}
\examples{
\dontrun{
# first run \link{compareNets} function example and pass the \code{cmpr_rslts} 

# plot permutation distribution:
 plotPermutes(cmpr_rslts)
}

}
